Added convolution example with PG1634
This example fits the image of PG1634+706 using a Beta2D model convolved with the corresponding PSF from the archive. For now, convolution and fitting are done in the same grid as the data, but I will next add the binning part that will allow us to fit when model and data are in different grids.